+++ pygount/write.py
@@ -4,11 +4,20 @@ Writers to store the results of a pygount analysis.
 # Copyright (c) 2016-2022, Thomas Aglassinger.
 # All rights reserved. Distributed under the BSD License.
 import datetime
+import json
 import math
 import os
 from xml.etree import ElementTree
 
+import pygount
+
+from . import SourceAnalysis
 from .summary import ProjectSummary
+
+#: Version of cloc the --format=cloc-xml pretends to be.
+CLOC_VERSION = "1.60"
+
+JSON_FORMAT_VERSION = "1.0.0"
 
 
 class BaseWriter:
@@ -19,12 +28,12 @@ class BaseWriter:
         except AttributeError:
             self.target_name = "<io>"
         self.project_summary = ProjectSummary()
-        self.start_time = datetime.datetime.now()
-        self.end_time = None
-        self.sources_per_second = 0
+        self.started_at = datetime.datetime.utcnow()
+        self.finished_at = None
+        self.files_per_second = 0
         self.lines_per_second = 0
         self.duration = None
-        self.duration_in_seconds = None
+        self.duration_in_seconds = 0.0
 
     def __enter__(self):
         return self
@@ -36,14 +45,14 @@ class BaseWriter:
         self.project_summary.add(source_analysis)
 
     def close(self):
-        self.end_time = datetime.datetime.now()
-        self.duration = self.end_time - self.start_time
+        self.finished_at = datetime.datetime.utcnow()
+        self.duration = self.finished_at - self.started_at
         self.duration_in_seconds = (
             (1e-6 * self.duration.microseconds) + self.duration.seconds + self.duration.days * 3600 * 24
         )
         if self.duration_in_seconds > 0:
             self.lines_per_second = self.project_summary.total_file_count / self.duration_in_seconds
-            self.sources_per_second = self.project_summary.total_file_count / self.duration_in_seconds
+            self.files_per_second = self.project_summary.total_file_count / self.duration_in_seconds
 
 
 class LineWriter(BaseWriter):
@@ -75,7 +84,7 @@ class ClocXmlWriter(BaseWriter):
         self._results_element = ElementTree.Element("results")
         self._header_element = ElementTree.SubElement(self._results_element, "header")
         ElementTree.SubElement(self._header_element, "cloc_url", text="https://github.com/roskakori/pygount")
-        ElementTree.SubElement(self._header_element, "cloc_version", text="1.60")
+        ElementTree.SubElement(self._header_element, "cloc_version", text=CLOC_VERSION)
         self._files_element = ElementTree.SubElement(self._results_element, "files")
 
     def __exit__(self, exc_type, exc_val, exc_tb):
@@ -83,11 +92,11 @@ class ClocXmlWriter(BaseWriter):
             # Only write the XML if everything works out.
             self.close()
 
-    def add(self, source_analysis):
+    def add(self, source_analysis: SourceAnalysis):
         super().add(source_analysis)
         file_attributes = {
             "blank": str(source_analysis.empty_count),
-            "code": str(source_analysis.code_count + source_analysis.string_count),
+            "code": str(source_analysis.source_count),
             "comment": str(source_analysis.documentation_count),
             "language": source_analysis.language,
             "name": source_analysis.path,
@@ -100,7 +109,7 @@ class ClocXmlWriter(BaseWriter):
         ElementTree.SubElement(self._header_element, "elapsed_seconds", text="%.f" % self.duration_in_seconds)
         ElementTree.SubElement(self._header_element, "n_files", text="%d" % self.project_summary.total_file_count)
         ElementTree.SubElement(self._header_element, "n_lines", text="%d" % self.project_summary.total_line_count)
-        ElementTree.SubElement(self._header_element, "files_per_second", text="%.f" % self.sources_per_second)
+        ElementTree.SubElement(self._header_element, "files_per_second", text="%.f" % self.files_per_second)
         ElementTree.SubElement(self._header_element, "lines_per_second", text="%.f" % self.lines_per_second)
         ElementTree.SubElement(self._header_element, "report_file", text=self.target_name)
 
@@ -267,6 +276,66 @@ class SummaryWriter(BaseWriter):
             percentage_width=percentage_width,
         )
         self._target_stream.write(summary_footer + os.linesep)
+
+
+class JsonWriter(BaseWriter):
+    """
+    Writer JSON output, ideal for further automatic processing.
+    """
+
+    def __init__(self, target_stream):
+        super().__init__(target_stream)
+        self.source_analyses = []
+
+    def add(self, source_analysis: SourceAnalysis):
+        super().add(source_analysis)
+        self.source_analyses.append(
+            {
+                "path": source_analysis.path,
+                "sourceCount": source_analysis.source_count,
+                "emptyCount": source_analysis.empty_count,
+                "documentationCount": source_analysis.documentation_count,
+                "group": source_analysis.group,
+                "isCountable": source_analysis.is_countable,
+                "language": source_analysis.language,
+                "state": source_analysis.state.name,
+                "stateInfo": source_analysis.state_info,
+            }
+        )
+
+    def close(self):
+        # NOTE: We are using camel case for naming here to follow JSLint's guidelines, see <https://www.jslint.com/>.
+        super().close()
+        json_map = {
+            "formatVersion": JSON_FORMAT_VERSION,
+            "pygountVersion": pygount.__version__,
+            "files": self.source_analyses,
+            "languages": [
+                {
+                    "documentationCount": language_summary.documentation_count,
+                    "emptyCount": language_summary.empty_count,
+                    "fileCount": language_summary.file_count,
+                    "isPseudoLanguage": language_summary.is_pseudo_language,
+                    "language": language_summary.language,
+                    "sourceCount": language_summary.source_count,
+                }
+                for language_summary in self.project_summary.language_to_language_summary_map.values()
+            ],
+            "runtime": {
+                "durationInSeconds": self.duration_in_seconds,
+                "filesPerSecond": self.files_per_second,
+                "finishedAt": self.finished_at.isoformat(),
+                "linesPerSecond": self.lines_per_second,
+                "startedAt": self.started_at.isoformat(),
+            },
+            "summary": {
+                "totalDocumentationCount": self.project_summary.total_documentation_count,
+                "totalEmptyCount": self.project_summary.total_empty_count,
+                "totalFileCount": self.project_summary.total_file_count,
+                "totalSourceCount": self.project_summary.total_source_count,
+            },
+        }
+        json.dump(json_map, self._target_stream)
 
 
 def digit_width(line_count):